requirements.txt
aiogram
sqlalchemy
aiosqlite
pytz
PyYAML==6.0.1